query-service/requirements.txt
fastapi
uvicorn
kafka-python
psycopg2-binary
pydantic